requirements.txt
# Core ML dependencies
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
albumentations>=1.3.1
opencv-python>=4.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Visualization and tracking
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.15.0
tqdm>=4.65.0

# Utilities
pycocotools>=2.0.7
jsonlines>=3.1.0

# Optimization
accelerate>=0.21.0